PCM Tones Demo
Got help from GPT, it is based on the code from the previous commits. It came up with this function to generate these scale-ish tones. It's not too musical yet I feel. Well, actually I like it a lot in a music sense. Maybe I can try to make something with this.

I am really curious about writing music with math now. Hope that makes sense. It's not for being "smart", but for using those technical concepts creatively and artistically. It's meant to express complexity and intricacies.