Web Setup
Okay this will be ready to be a web app real soon to be.

For preface, the last few commits I made the type changes with types/node installed temporarily. It was just to migrate the code properly with type safety. I don't intend to make this a Node-specific project, next it will be browser-first (hopefully).