PCM from Plugins
The concept here is that the PCM should be what the plugin generates for us, rather than the data it is produced from.

I think I might be mixing things up here, going to look into it. Should I be generating random PCM, or should I make random/procedural data, which is then turned into PCM?